feat(resume-parser): chnage resume parser to version 1
